Task_1_A/requirements.txt
apache-airflow==2.7.0
mlflow==2.7.0
scikit-learn==1.3.0
flask==2.2.5
Flask-Session==0.4.0
psycopg2-binary==2.9.7
pandas==1.5.3
joblib==1.3.2
pendulum==2.1.2
